\name{calibHiC}
\alias{calibHiC}
\docType{data}
\title{
in-silico 100*100 calibrated Hi-C contact matrix 
}
\description{
Generated with the function \code{\link{calibrate}} given the in-silico Hi-C contact matrix \code{\link{hic}}. 
}
\usage{data(calibHiC)}

\details{
This data structure is used to illustrate the usage of \code{\link{lsmacof}}.
}

\references{
Y. Shavit, F.K. Hamey, P. Lio', FisHiCal: an R package for iterative FISH-based calibration of Hi-C data, 2014 (submitted).
}

\keyword{datasets}
