\name{conf}
\alias{conf}
\docType{data}
\title{
A 3D random configuration of 100 points
}
\description{
A numeric 100*3 matrix giving the 3D coordinates for each point. The configuration was generated as described in Shavit, Hamey and Lio' (2014), as part of an in-silico evaluation.
}
\usage{data(conf)}

\references{
Y. Shavit, F.K. Hamey, P. Lio', FisHiCal: an R package for iterative FISH-based calibration of Hi-C data, 2014 (submitted).
}
\keyword{datasets}
