\name{spatialInc}
\alias{spatialInc}
\docType{data}
\title{
Spatial inconsistencies detected for loci in fibroblasts chromosome 1
}
\description{
The list of spatial inconsistencies detected by \code{\link{searchInc}} for a fibroblasts chromosome 1.
}
\usage{data(spatialInc)}

\source{
Data generated as described in Shavit, Hamey and Lio' (2014) (submitted) and used to iilustrate the usage of \code{\link{plotInc}} and \code{\link{summaryInc}}.
}
\references{
Y. Shavit, F.K. Hamey, P. Lio', FisHiCal: an R package for iterative FISH-based calibration of Hi-C data, 2014 (submitted).
}

\keyword{datasets}
